+++ angular-workspace/projects/trimble-oss/modus-angular-components/src/lib/stencil-generated/index.ts
@@ -11,6 +11,7 @@ export const DIRECTIVES = [
   d.ModusCard,
   d.ModusCheckbox,
   d.ModusChip,
+  d.ModusDataTable,
   d.ModusDropdown,
   d.ModusFileDropzone,
   d.ModusList,